+++ src/test/workspace/js/syntax.error.js
@@ -1,0 +1,7 @@
+const a = 'aaaa';
+let b = 'bbbbb;
+var c = 'ccccccccc";
+
+for (let i = 0; i < 100; i++) {
+    console.log(i);
+}